refacto: change end line type from Windows to Linux type
