+++ src/engine/registry/Registry.cpp
@@ -24,7 +24,6 @@ void ES::Engine::Registry::RunSystems()
     }
 }
 
-
 bool ES::Engine::Registry::IsEntityValid(entt::entity entity)
 {
     return GetRegistry().valid(entity);
